mtf_be/requirement.txt
Django>=4.0
djangorestframework
gunicorn
django-cors-headers
psycopg2-binary
dj-database-url
yfinance